Reformat stand-alone footnote.
In asciidoc 8.2.7 the footnote macro expands to "[n]" in a superscript, where n
is automatically generated.  In this case, we don't want that.  Ideally we'd
like the footnote text included in the footnote section, but not as a numbered
footnote.  This isn't possible, so I introduce a section just before the
footnotes and place the text there, in the normal flow.